Add response class to store + format HTTP response
The response class is used to store header information and body
information relating to a certain request to the server. The API for
this class includes the response.addResponse(result, data) where result is
an enum describing if this is a valid HTTP request or not and data is a
std::string containing the entire request to the server. Once
addResponse() is run, response.getResponse() will return a vector of
Boost const_buffer where the first buffer is the response header and the
second contains the response body.

IMPORTANT: the underlying data kept inside each of the returned buffers
is handled by the response class, so do not remove any response object
until after the response has been sent and the async_write callback
function is entered. Otherwise, there will be undefined behavior.

TODO: More HTTP header information should be added into the response
like the date and server name. We also need more tests for the response
class.

Change-Id: Ib64a7e3062cc00452b009c3cba4ea53c583c561b