Remove connection dependencies from session class
Created a new connection class which performs the reading and writing to
the sockets. Now, the session class delegates control over reading and
writting to the connection class which is created by the server class.
The next step is to remove all these dependecies from the server class
and place them into server_main.cc.

I chose to use smart pointers to deal with dynamic memory because that
will make keeping track of the memory much simpler.

In tests/mock_connection.cc I created a class which implements a very
basic mock connection which can be used for the other unit tests. A
connection is now an interface and tcp_connection is an implementation
of that interface for tcp connections. The connection interface is not
perfect because it still relies on a tcp socket instead of a generalized
socket.

I also renamed the functions inside of session to make it more readable.
handle_read is now received_req, and handle_write is now wait_for_req.

Change-Id: I8a007bf12743bb1d0c4b852d178cc0ed98669783