Refactor for smart pointers and update response.cc
To encode the concept of object ownership, I used smart pointers
whenever possible. Now, a server object creates sessions whenever a new
connection is established. This session will stay valid as long as there
is a coroutine which is holding a pointer to it, and multiple objects
can control ownership of a single session. A session is given a socket
which it controls and then uses a connection object to read and write to
this socket.

I updated the unit tests to fit the new class implementation and running
with Valgrind produces 0 memory leaks.

Change-Id: I6d01ca2071cef2532837325afcebc525b3961945