Fix indendation inconsistency and remove logs
Some of the files (mostly config and request parser) were not matching
our coding style. This commit will update them to match our style. Also,
I removed the log/* file from the repo and added the log directory to
the gitignore so that we do not accidentally push them to the repo.

Change-Id: Id522b4661e01bb527b4320ee40c0ddae78c64a13