Add 404page into deploy image
This change fixes the 404 error system to only send a 400 error with a
request was received which could not be parsed. The 404 error is sent
whenever a request is made to a destination which does not have any
request handler setup for it, or if the static handler for a certain
prefix was unable to find the requested file.

Change-Id: I1d18e237dba8cf94ef4e541ce31c1ab673731223