Update request and response to common API (almost)
The request parser is now greatly simplified to just meet the needs of
our simple HTTP parsing. The previous request parser was parsing many
different fields which are no longer necessary due to the new common
API.

This update to the request handler API is not complete since I must
still remove all the methods from the response class since that class is
not longer allowed to have any methods. I will used a response builder
class which I will refactor all the methods to.

Change-Id: I455a66f999d6f3624c0d92f98a88cbfbbfb0b326