Remove methods from response class, use builder
All the methods from the response class are not moved to
response_builder which is used whenever we want to create a response.
This will allow us to create common responses like
response_builder.make_404() and then finally build the
boost::const_buffer to use in order to send the response. It was
important to keep unique_ptrs to the final response string generated by
the build() method since the boost buffers do not keep the memory alive.
I ran into problems with the string going out of scope even though the
boost buffer was still in scope.

Change-Id: Iddd6072da8e2d2a9205d3240a44d9545b7feb6a2