Make each request get its own thread
The request parser now saves state so that between two separate reads in
a session we can be reading the same request and incrementally build our
final request. This will allow us to send half a request, and then the
other half later if we want to.

For multithreading, the requests in a session are no longer given their
own thread because this is overkill and makes processing requests much
harder. Two different connections (sessions) will be able to run in
parallel since they use different sockets and are not accessing any
shared memory other than the logger which is synchronized with mutexes.

For testing, the session tests are updated to remove unecessary tests
and included some corner cases which we missed before.

Change-Id: I1dfb3a293e7ff4885925a4275bf0dceee0e03331