Reverse Proxy Handler
Hey Team WNZA, this commit implements functionality of the reverse proxy handler. It receives a request,
formats it, forwards to the specified destination of the config, and retrieves/formats the response to send
back to the client. It also handles redirects.

#WHAT: Added the reverse proxy handler in team WNZA's web server

#WHY: We want to handle proxy requests -> receive a request on our server, send it to a different
one, retrieve the response, and send the response back to the client.

#HOW: Added a config block in /tests/test_server_config that specifies 'dest', 'port' is also
configurable but has a default value of 80. Added necessary dependencies across files and into
CMakeLists.txt. Finally implemented all virtual methods of request_handler, the class from which
our proxy request handler inherits from. Utilized helper methods to simplify the logic of handle_request.

#TESTING: No new testing for the proxy request handler.
  - All previous tests still pass
  - Behavior verified via command line (curl) and docker build (Chrome browser)

Note: logging will come in the next commit

Change-Id: If7c277b91c318c5a787c4fe9ff4e07ea8710665a