Reverse Proxy Integration Test
What: Implementing an integration test to ensure that the reverse
proxying functionality is implemented correctly.

Why: Integration is needed to verify interactions between two
servers resolve in the right way.

How: Extending the integration test by adding an additional section
which has a new server that uses a different config to proxy.

Tests: Pass, coverage is 91% and line coverage is 53%
Change-Id: I633ad215e5f79f975a1c0a17aa79700765ae6629