Drop connections if there are already 100 open
I keep an atomic counter which is passed to all the connections and
incremented on connection creation and decremented on connection
deletion. If we try to create a connection when the count is 100, then
the server waits for 3 seconds before initiating any new connections.
This wait time is arbitrary and we may want to try different values to
wait for. This does not change any other parts of the server, so no
tests change and no new tests are created because we must load test to
actually test this feature.

Change-Id: I1b7859aacc0a149c04a07df49a24dbb185b66cd7