Use a thread pool to manage these threads dynamically
Now the multithreading is optimized using the method proposed by Zongze and I followed the previous logic in the test when testing them.

All of the test are passed, and I also uploaded the sample_log_0.log so that you could check the result of the 16 threads in my testing.

The 16 different threadIDs is recorded in the log file.

Change-Id: Id2e79d053f45765350910b429ae8095165227d43